updated to rev 10488

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/dataelement/comparator/DataElementCategoryComboSizeComparator.java
@@ -39,6 +39,6 @@ public class DataElementCategoryComboSizeComparator
 {
     public int compare( DataElementCategoryCombo o1, DataElementCategoryCombo o2 )
     {
-        return (int) (o1.getOptionCombos().size() - o2.getOptionCombos().size());
+        return o1.getOptionCombos().size() - o2.getOptionCombos().size();
     }
 }
